requirements.txt
pandas==1.2.1
requests==2.25.1
streamlit==0.83.0
facebook-scraper==0.2.43
Pillow==8.1.0